Release/v0.1.0

+++ contracts/libraries/Transfers.sol
@@ -9,10 +9,13 @@ library Transfers {
     /// @param  to      The recipient of the ERC20 token
     /// @param  value   The amount of ERC20 to transfer
     /// @return Whether or not the call was successful
-    function safeTransfer(IERC20 token, address to, uint value) internal returns (bool) {
-        (bool success, bytes memory returnData) = address(token).call(
-            abi.encodeWithSelector(token.transfer.selector, to, value)
-        );
+    function safeTransfer(
+        IERC20 token,
+        address to,
+        uint256 value
+    ) internal returns (bool) {
+        (bool success, bytes memory returnData) =
+            address(token).call(abi.encodeWithSelector(token.transfer.selector, to, value));
         require(success && (returnData.length == 0 || abi.decode(returnData, (bool))), "Transfer fail");
         return success;
     }
